backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
geoalchemy2>=0.14.2
redis>=5.0.1
celery>=5.3.4
pydantic>=2.5.0
pydantic-settings>=2.1.0
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.2.0
requests>=2.31.0
httpx>=0.25.2
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
boto3>=1.34.0
geopy>=2.4.0
websockets>=12.0
python-socketio>=5.10.0
aiofiles>=23.2.1
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0
coverage>=7.3.2
pre-commit>=3.6.0
python-dotenv>=1.0.0
numpy>=1.26.0
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
